debug

+++ util.py
@@ -62,7 +62,11 @@ def save_phantom_output(output, pbconf):
 
     output_filename = os.path.join(os.environ['PHANTOM_DATA'], pbconf['name'], 'phantom_output')
 
-    # if os.path.exists(output_filename):
-    with open(output_filename, 'w') as f:
-        # pickle.load(f)
-        f.write(output)
+    if os.path.exists(output_filename):
+        with open(output_filename, 'a') as f:
+            # pickle.load(f)
+            f.write(output)
+
+    else:
+        with open(output_filename, 'w') as f:
+            f.write(output)
